+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - Improve error messages. Add --reeask-version
 - Use ggwsm in pipelines
 
+### Fixed
+
+- Fix issues in gg
+
 ## 4.1.0 - 2026-08-14
 
 ### Changed