+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 # Changelog
 
-## Unreleased
+## 4.1.1 - 2026-08-18
 
 ### Changed
 
+- Improve error messages. Add --reeask-version
 - Use ggwsm in pipelines
+
+### Fixed
+
+- Fix issues in gg
 
 ## 4.1.0 - 2026-08-14